# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Surge API",
    "description": "Surge is an evidence-first investigation and response agent.\n\nCreate an investigation, subscribe to its Server-Sent Events stream, inspect evidence, hypotheses and\nclaims, approve the proposed action, and read back the independently verified outcome. Evaluation\nendpoints expose reliability metrics computed from real scenario runs.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/investigations": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Create Investigation",
        "operationId": "create_investigation_api_investigations_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvestigationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "List Investigations",
        "operationId": "list_investigations_api_investigations_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated statuses",
              "title": "Status"
            },
            "description": "Comma-separated statuses"
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only investigations created with this session_id",
              "title": "Session Id"
            },
            "description": "Only investigations created with this session_id"
          },
          {
            "name": "include_evaluation",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Evaluation"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Investigation",
        "operationId": "get_investigation_api_investigations__investigation_id__get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/events": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Events",
        "operationId": "get_events_api_investigations__investigation_id__events_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventOut"
                  },
                  "title": "Response Get Events Api Investigations  Investigation Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/evidence": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Evidence",
        "operationId": "get_evidence_api_investigations__investigation_id__evidence_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvidenceOut"
                  },
                  "title": "Response Get Evidence Api Investigations  Investigation Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/hypotheses": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Hypotheses",
        "operationId": "get_hypotheses_api_investigations__investigation_id__hypotheses_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HypothesisOut"
                  },
                  "title": "Response Get Hypotheses Api Investigations  Investigation Id  Hypotheses Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/actions": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Actions",
        "operationId": "get_actions_api_investigations__investigation_id__actions_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActionOut"
                  },
                  "title": "Response Get Actions Api Investigations  Investigation Id  Actions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/claims": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Get Claims",
        "operationId": "get_claims_api_investigations__investigation_id__claims_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClaimOut"
                  },
                  "title": "Response Get Claims Api Investigations  Investigation Id  Claims Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/approve": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Approve Action",
        "operationId": "approve_action_api_investigations__investigation_id__approve_post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/cancel": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Cancel Investigation",
        "operationId": "cancel_investigation_api_investigations__investigation_id__cancel_post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Investigation Api Investigations  Investigation Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/rerun": {
      "post": {
        "tags": [
          "investigations"
        ],
        "summary": "Rerun Investigation",
        "operationId": "rerun_investigation_api_investigations__investigation_id__rerun_post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RerunRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationCreated"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/stream": {
      "get": {
        "tags": [
          "investigations"
        ],
        "summary": "Stream Events",
        "operationId": "stream_events_api_investigations__investigation_id__stream_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "After"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-Sent Events of EventOut objects",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApprovalMode": {
        "type": "string",
        "enum": [
          "required",
          "auto_low_risk"
        ],
        "title": "ApprovalMode"
      },
...
      "CreateInvestigationRequest": {
        "properties": {
          "request": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 5,
            "title": "Request",
            "description": "Natural-language description of the problem"
          },
          "time_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TimeRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "mode": {
            "type": "string",
            "const": "interactive",
            "title": "Mode",
            "default": "interactive"
          },
          "approval_mode": {
            "$ref": "#/components/schemas/ApprovalMode",
            "default": "required"
          },
          "connector_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "DEMO",
                  "REAL"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Connector Mode",
            "description": "Override the server's default connector mode for every app"
          },
          "disabled_apps": {
            "items": {
              "$ref": "#/components/schemas/App"
            },
            "type": "array",
            "title": "Disabled Apps"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z0-9_]{3,80}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id",
            "description": "Seeded demo world to use"
          },
          "inject_failures": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "title": "Inject Failures",
            "description": "Fault presets, see /api/connectors"
          },
          "reasoning": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReasoningMode"
              },
              {
                "type": "null"
              }
            ],
            "description": "llm or heuristic; defaults to server setting"
          },
          "persona": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(founder|student|engineer|designer|marketer|other)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persona",
            "description": "Overrides the signed-in user's saved persona for this one investigation; ignored if not signed in and no persona given"
          },
          "session_id": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[\\w:.\\-]+$",
            "title": "Session Id",
            "default": "anonymous"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request"
        ],
        "title": "CreateInvestigationRequest"
      },
...
      "InvestigationCreated": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stream_url": {
            "type": "string",
            "title": "Stream Url"
          },
          "connector_mode": {
            "type": "string",
            "title": "Connector Mode"
          },
          "reasoning_mode": {
            "type": "string",
            "title": "Reasoning Mode"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "injected_faults": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Injected Faults"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "stream_url",
          "connector_mode",
          "reasoning_mode",
          "scenario_id",
          "injected_faults"
        ],
        "title": "InvestigationCreated"
      },
...
    }
  }
}